Cleanup prior; add remote_file_cache.py
